+++ vcfpy/header.py
@@ -301,8 +301,9 @@ class Header:
         :param mapping: ``OrderedDict`` with mapping to add.  It is
             recommended to use ``OrderedDict`` over ``dict`` as this makes
             the result reproducible
+        :return: ``False`` on conflicting line and ``True`` otherwise
         """
-        self.add_line(FilterHeaderLine.from_mapping(mapping))
+        return self.add_line(FilterHeaderLine.from_mapping(mapping))
 
     def add_contig_line(self, mapping):
         """Add "contig" header line constructed from the given mapping
@@ -310,8 +311,9 @@ class Header:
         :param mapping: ``OrderedDict`` with mapping to add.  It is
             recommended to use ``OrderedDict`` over ``dict`` as this makes
             the result reproducible
+        :return: ``False`` on conflicting line and ``True`` otherwise
         """
-        self.add_line(ContigHeaderLine.from_mapping(mapping))
+        return self.add_line(ContigHeaderLine.from_mapping(mapping))
 
     def add_info_line(self, mapping):
         """Add INFO header line constructed from the given mapping
@@ -319,8 +321,9 @@ class Header:
         :param mapping: ``OrderedDict`` with mapping to add.  It is
             recommended to use ``OrderedDict`` over ``dict`` as this makes
             the result reproducible
+        :return: ``False`` on conflicting line and ``True`` otherwise
         """
-        self.add_line(InfoHeaderLine.from_mapping(mapping))
+        return self.add_line(InfoHeaderLine.from_mapping(mapping))
 
     def add_format_line(self, mapping):
         """Add FORMAT header line constructed from the given mapping
@@ -328,8 +331,9 @@ class Header:
         :param mapping: ``OrderedDict`` with mapping to add.  It is
             recommended to use ``OrderedDict`` over ``dict`` as this makes
             the result reproducible
+        :return: ``False`` on conflicting line and ``True`` otherwise
         """
-        self.add_line(FormatHeaderLine.from_mapping(mapping))
+        return self.add_line(FormatHeaderLine.from_mapping(mapping))
 
     def format_ids(self):
         """Return list of all format IDs"""
@@ -351,7 +355,10 @@ class Header:
             return []
 
     def add_line(self, header_line):
-        """Add header line, updating any necessary support indices"""
+        """Add header line, updating any necessary support indices
+
+        :return: ``False`` on conflicting line and ``True`` otherwise
+        """
         self.lines.append(header_line)
         self._indices.setdefault(header_line.key, OrderedDict())
         if not hasattr(header_line, 'mapping'):
@@ -361,9 +368,11 @@ class Header:
                 ('Detected duplicate header line with type {} and ID {}. '
                  'Ignoring this and subsequent one').format(
                      header_line.key, header_line.mapping['ID']))
+            return False
         else:
             self._indices[header_line.key][
                 header_line.mapping['ID']] = header_line
+            return True
 
     def get_info_field_info(self, key):
         """Return :py:class:`FieldInfo` for the given INFO field"""