+++ HISTORY.rst
@@ -5,6 +5,7 @@ History
 HEAD
 ----
 
+* ``Header.add_*_line`` return a ``bool`` no indicating any conflicts
 * Construction of Writer uses samples within header and no extra parameter (breaks API)
 
 0.7.0 (2016-09-25)